third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/word-spacing-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'word-spacing' to the 'normal' keyword: normal
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'word-spacing' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'word-spacing' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'word-spacing' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'word-spacing' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
Harness: the test ran to completion.
